cyrus-ai/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
chromadb>=0.5.23
numpy>=1.26.4,<2
torch>=2.2.0
transformers>=4.44.2
accelerate>=0.33.0
datasets>=2.21.0
sentencepiece>=0.2.0
pydantic>=2.6.0
openai
redis
tiktoken
aioredis
loguru>=0.7.2
prometheus-client>=0.20.0
requests>=2.31.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp>=1.24.0
opentelemetry-instrumentation-fastapi>=0.45b0
pymavlink>=2.4.41
opencv-python-headless>=4.9.0.80